backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-genai